Bitcoin_trader/requirements.txt
fastapi
uvicorn
pandas
numpy==1.26.4
scikit-learn
tensorflow
keras
pydantic
pandas_ta
xgboost
tabulate
python-binance
python-dotenv